added Typescript and more game state
